Trekk ut visningsmodal til eigen komponent
