+++ src/minoversikt/huskelapp/HuskelappIkonInngang.tsx
@@ -1,24 +1,20 @@
 import React, {useState} from 'react';
+import {Button} from '@navikt/ds-react';
 import {BrukerModell, HuskelappModell} from '../../model-interfaces';
 import {LagEllerEndreHuskelappModal} from './redigering/LagEllerEndreHuskelappModal';
-import {Huskelapp} from './Huskelapp';
-import {TrashIcon} from '@navikt/aksel-icons';
-import {handleSlettHuskelapp} from './redigering/slettHuskelapp';
-import {ThunkDispatch} from 'redux-thunk';
-import {AppState} from '../../reducer';
-import {AnyAction} from 'redux';
-import {useDispatch} from 'react-redux';
-import {usePortefoljeSelector} from '../../hooks/redux/use-portefolje-selector';
-import {OversiktType} from '../../ducks/ui/listevisning';
 import {ReactComponent as HuskelappIkon} from '../../components/ikoner/huskelapp/huskelapp.svg';
 import {ReactComponent as HuskelappIkonTomt} from '../../components/ikoner/huskelapp/huskelapp_stiplet.svg';
-import {Button, Heading, Modal} from '@navikt/ds-react';
+import {HuskelappModal} from './modalvisning/HuskelappModal';
 
 export const HuskelappIkonInngang = ({bruker}: {bruker: BrukerModell}) => {
-    const dispatch: ThunkDispatch<AppState, any, AnyAction> = useDispatch();
-    const {enhetId} = usePortefoljeSelector(OversiktType.minOversikt);
     const [modalLagEllerEndreHuskelappSkalVises, setModalLagEllerEndreHuskelappSkalVises] = useState<boolean>(false);
     const [modalVisHuskelappSkalVises, setModalVisHuskelappSkalVises] = useState<boolean>(false);
+
+    function redigerHuskelapp() {
+        setModalVisHuskelappSkalVises(false);
+        setModalLagEllerEndreHuskelappSkalVises(true);
+    }
+
     return (
         <>
             <Button
@@ -52,48 +48,13 @@ export const HuskelappIkonInngang = ({bruker}: {bruker: BrukerModell}) => {
                 />
             )}
             {modalVisHuskelappSkalVises && (
-                <Modal
+                <HuskelappModal
                     open={modalVisHuskelappSkalVises}
                     onClose={() => setModalVisHuskelappSkalVises(false)}
-                    closeOnBackdropClick={true}
-                >
-                    <Modal.Header>
-                        <Heading size="medium" level="1" spacing className="huskelapp-modal__header">
-                            <HuskelappIkon />
-                            Huskelapp
-                        </Heading>
-                    </Modal.Header>
-                    <Modal.Body>
-                        <Huskelapp huskelapp={bruker.huskelapp!!} className="huskelapp-i-modal" />
-                        <div className="huskelapp-handlingsknapper">
-                            <Button
-                                type="button"
-                                size="xsmall"
-                                variant="secondary"
-                                onClick={() => {
-                                    //todo varsel modal på at det kommer til å slettes
-                                    handleSlettHuskelapp(dispatch, bruker.huskelapp!!, bruker.fnr, enhetId!!).then(() =>
-                                        setModalVisHuskelappSkalVises(false)
-                                    );
-                                }}
-                                icon={<TrashIcon />}
-                            >
-                                Slett
-                            </Button>
-                            <Button
-                                type="button"
-                                size="xsmall"
-                                variant="primary"
-                                onClick={() => {
-                                    setModalVisHuskelappSkalVises(false);
-                                    setModalLagEllerEndreHuskelappSkalVises(true);
-                                }}
-                            >
-                                Endre
-                            </Button>
-                        </div>
-                    </Modal.Body>
-                </Modal>
+                    bruker={bruker}
+                    redigerHuskelapp={redigerHuskelapp}
+                    setModalVisHuskelappSkalVises={setModalVisHuskelappSkalVises}
+                />
             )}
         </>
     );
